requirements.txt
Flask == 2.0.2
Flask-SQLAlchemy == 2.5.1
flask-restx == 0.5.1
marshmallow == 3.14.1
SQLAlchemy == 1.3.23
marshmallow-sqlalchemy == 0.27.0
flask_jwt_extended == 4.3.1
python-decouple == 3.5
Flask-Mail-SendGrid == 0.3